+++ src/packageAgent.ts
@@ -1,12 +1,6 @@
-import { createAgent } from '@veramo/core';
-import { DIDManager, MemoryDIDStore } from '@veramo/did-manager';
-import { KeyManager, MemoryKeyStore, MemoryPrivateKeyStore } from '@veramo/key-manager';
-import { CredentialPlugin } from '@veramo/credential-w3c';
-import { KeyManagementSystem } from '@veramo/kms-local';
-import { getUniversalResolverFor, DIDResolverPlugin } from '@veramo/did-resolver';
-import { KeyDIDProvider } from '@veramo/did-provider-key';
+import { createOVAgent, createCheqdProvider, CheqdNetwork, keyStore, privateKeyStore, AgentStore } from './OVAgent';
+import { getUniversalResolverFor } from '@veramo/did-resolver';
 import { CheqdDIDProvider } from '@cheqd/did-provider-cheqd';
-import { Resolver } from 'did-resolver';
 import { IOVAgent, ICreateVerifiableCredentialArgs, ManagedKeyInfo, DIDAssertionCredential, VerifiableCredential, IIdentifier } from '@originvault/ov-types';
 import { getSelfBundlePrivateKey, getPackageDIDFromPackageJson, getSelfBundleHash } from './packageManager.js';
 import { generateDIDKey } from './didKey.js';
@@ -18,26 +12,19 @@ import { createResource } from './resourceManager.js';
 import { getEnvironmentMetadata } from './environment.js';
 import path from 'path';
 import { fileURLToPath } from 'url';
-import { KeyringPair$Json, KeyringPair$Meta } from '@polkadot/keyring/types.js';
 import { co2 } from "@tgwf/co2";
+import { getResolver } from "@verida/vda-did-resolver";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
 
 dotenv.config();
 
-const universalResolver = getUniversalResolverFor(['cheqd', 'key']);
-
-// Create a key store instance
-const keyStore = new MemoryKeyStore();
-const privateKeyStore = new MemoryPrivateKeyStore();
-
-declare enum CheqdNetwork {
-    Mainnet = "mainnet",
-    Testnet = "testnet"
-}
+const universalResolver = getUniversalResolverFor(['cheqd', 'key', 'vda']);
+const vdaResolver = getResolver();
 
 const packageJsonPath = path.join(__dirname, '../package.json');
+
 let cheqdMainnetProvider: CheqdDIDProvider | null = null;
 let packageAgent: IOVAgent | null = null;
 let currentDIDKey: string | null = null;
@@ -51,53 +38,26 @@ const initializePackageAgent = async ({ payerSeed, didRecoveryPhrase }: { payerS
     let cosmosPayerSeed = payerSeed || process.env.COSMOS_PAYER_SEED || '';
     let didMnemonic = didRecoveryPhrase || process.env.PACKAGE_DID_RECOVERY_PHRASE || '';
 
-    cheqdMainnetProvider = new CheqdDIDProvider({
-        defaultKms: 'local',
-        networkType: 'mainnet' as CheqdNetwork,
-        dkgOptions: { chain: 'cheqdMainnet' },
-        rpcUrl: process.env.CHEQD_RPC_URL || 'https://cheqd.originvault.box:443',
-        cosmosPayerSeed,
-    })
+    cheqdMainnetProvider = createCheqdProvider(CheqdNetwork.Mainnet, cosmosPayerSeed, process.env.CHEQD_RPC_URL || 'https://cheqd.originvault.box:443');
 
-    packageAgent = createAgent<IOVAgent>({
-        plugins: [
-            new KeyManager({
-                store: keyStore,
-                kms: {
-                    local: new KeyManagementSystem(privateKeyStore),
-                },
-            }),
-            new DIDManager({
-                store: new MemoryDIDStore(),
-                defaultProvider: 'did:cheqd:mainnet',
-                providers: {
-                    'did:cheqd:mainnet': cheqdMainnetProvider,
-                    'did:cheqd:testnet': new CheqdDIDProvider({
-                        defaultKms: 'local',
-                        networkType: 'testnet' as CheqdNetwork,
-                        dkgOptions: { chain: 'cheqdTestnet' },
-                        rpcUrl: process.env.CHEQD_RPC_URL || 'https://cheqd.originvault.box:443',
-                        cosmosPayerSeed,
-                    }),
-                    'did:key': new KeyDIDProvider({
-                        defaultKms: 'local',
-                    }),
-                }
-            }),
-            new DIDResolverPlugin({
-                resolver: new Resolver(universalResolver)
-            }),
-            new CredentialPlugin(),
-        ],
-    })
+    packageAgent = createOVAgent(cheqdMainnetProvider, universalResolver, vdaResolver);
+
+    if(!packageAgent) {
+        throw new Error("Package agent could not be initialized");
+    }
 
     const packageJsonDIDString = await getPackageDIDFromPackageJson();
 
     if (didMnemonic) {
         const packagePrivateKey = await convertRecoveryToPrivateKey(didMnemonic);
-        const { credentials } = await importDID(packageJsonDIDString, packagePrivateKey, 'cheqd', packageAgent);
 
-        signedVCs.concat(credentials);
+        try {
+            const { credentials } = await importDID({ didString: packageJsonDIDString, privateKey: packagePrivateKey, method: 'cheqd', agent: packageAgent });
+            signedVCs.concat(credentials);
+        } catch (error) {
+            console.error("❌ Error importing DID:", error);
+            throw error;
+        }
     }
     
     // Generate did:web after agent initialization
@@ -251,20 +211,7 @@ const initializePackageAgent = async ({ payerSeed, didRecoveryPhrase }: { payerS
         return result;
     }
 
-    return { agent: packageAgent, did: packageJsonDIDString, key: currentDIDKey, credentials: signedVCs, publishWorkingKey, publishRelease };
-}
-
-interface AgentStore {
-    initialize: (args: { payerSeed?: string, didRecoveryPhrase?: string }) => Promise<{ agent: IOVAgent, did: string, key: string, credentials: VerifiableCredential[], publishWorkingKey: (() => Promise<string | undefined>) | null, publishRelease: (releaseCredential: any, name: string, version: string) => Promise<string | undefined> }>,
-    agent: IOVAgent | null,
-    keyStore: MemoryKeyStore,
-    cheqdMainnetProvider: CheqdDIDProvider | null,
-    listDids: (provider?: string) => Promise<IIdentifier[]>,
-    getDID: (didString: string) => Promise<KeyringPair$Meta | undefined>,
-    createDID: (props: { method: string, alias: string, isPrimary: boolean }) => Promise<{ did: IIdentifier, mnemonic: string, credentials: VerifiableCredential[] }>,
-    importDID: (didString: string, privateKey: string, method: string) => Promise<{ did: IIdentifier, credentials: VerifiableCredential[] }>,
-    getPrimaryDID: () => Promise<string>,
-    [key: string]: any,
+    return { agent: packageAgent, did: packageJsonDIDString, key: currentDIDKey, credentials: signedVCs, publishWorkingKey, publishRelease, privateKeyStore };
 }
 
 const packageStore: AgentStore = {
@@ -274,14 +221,15 @@ const packageStore: AgentStore = {
     cheqdMainnetProvider,
     listDids: async (provider?: string) => packageAgent ? listDIDs(packageAgent, provider) : [] as IIdentifier[],
     getDID: async (didString: string) => getDIDKeys(didString),
-    createDID: (props: { method: string, alias: string, isPrimary: boolean }) => packageAgent ? createDID({ ...props, agent: packageAgent }) : Promise.reject(new Error("Package agent not initialized")),
-    importDID: (didString: string, privateKey: string, method: string) => packageAgent ? importDID(didString, privateKey, method, packageAgent) : Promise.reject(new Error("Package agent not initialized")),
+    createDID: (props: { method: string, alias: string, isPrimary?: boolean }) => packageAgent ? createDID({ ...props, agent: packageAgent }) : Promise.reject(new Error("Package agent not initialized")),
+    importDID: (didString: string, privateKey: string, method: string) => packageAgent ? importDID({ didString, privateKey, method, agent: packageAgent }) : Promise.reject(new Error("Package agent not initialized")),
     getPrimaryDID: async () => await getPackageDIDFromPackageJson(),
     getBundleHash: async () => await getSelfBundleHash(),
     publishWorkingKey,
     publishRelease,
     didKey: currentDIDKey,
-    packageJsonPath
+    packageJsonPath,
+    privateKeyStore
 }
 
 export { packageStore };